Django/01_api_project/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-extensions==3.0.9
django-seed==0.2.2
djangorestframework==3.12.1
Faker==4.1.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
